Added NFC Reader // Refactored Discord
Implemented NFC Reader, Refactored Discord into a new file.
