+++ src/main.py
@@ -13,6 +13,9 @@ from tinkerforge.bricklet_piezo_speaker_v2 import BrickletPiezoSpeakerV2
 from tinkerforge.bricklet_ambient_light_v3 import BrickletAmbientLightV3
 from tinkerforge.bricklet_humidity_v2 import BrickletHumidityV2
 from tinkerforge.bricklet_e_paper_296x128 import BrickletEPaper296x128
+from tinkerforge.bricklet_nfc import BrickletNFC
+
+from discord import send
 
 import time
 
@@ -74,30 +77,7 @@ class SensorData:
 with open("wh.dat") as f:
     WEBHOOK = f.readline()
 
-class Discord:
-    # get the connection and make the request
-    host = "discord.com"
-    connection = http.client.HTTPSConnection(host)
 
-    @staticmethod
-    def send(message):
-        global WEBHOOK
-        # your webhook URL
-
-        payload = json.dumps({"content": message})
-
-        headers = {
-            "Content-Type": "application/json"
-        }
-
-        Discord.connection.request("POST", WEBHOOK, body=payload, headers=headers)
-
-        # get the response
-        response = Discord.connection.getresponse()
-        result = response.read()
-
-        # return back to the calling function with the result
-        return f"{response.status} {response.reason}\n{result.decode()}"
 
 IP = "172.20.10.242"
 PORT = 4223
@@ -116,6 +96,19 @@ def ambient_light_callback(illuminance):
 def moisture_callback(moisture):
     SENSOR_DATA.moisture.set_current(moisture / 100)
 
+def cb_reader_state_changed(state, idle, nfc):
+    if state == nfc.READER_STATE_REQUEST_TAG_ID_READY:
+        ret = nfc.reader_get_tag_id()
+
+        print("Found tag of type " +
+              str(ret.tag_type) +
+              " with ID [" +
+              " ".join(map(str, map('0x{:02X}'.format, ret.tag_id))) +
+              "]")
+
+    if idle:
+        nfc.reader_request_tag_id()
+
 if __name__ == "__main__":
     conn = IPConnection()
 
@@ -127,8 +120,10 @@ if __name__ == "__main__":
     ambient_light = BrickletAmbientLightV3("Pdw", conn)
     temp = BrickletPTCV2("Wcg", conn)
     moisture_sensore = BrickletHumidityV2("ViW", conn)
+    nfc = BrickletNFC("22ND", conn)
 
     conn.connect(IP, PORT)
+
 
     # register callbacks
     temp.register_callback(temp.CALLBACK_TEMPERATURE, temperature_callback)
@@ -140,6 +135,10 @@ if __name__ == "__main__":
     moisture_sensore.register_callback(moisture_sensore.CALLBACK_HUMIDITY, moisture_callback)
     moisture_sensore.set_humidity_callback_configuration(1000, False, "x", 0, 0)
 
+    nfc.register_callback(nfc.CALLBACK_READER_STATE_CHANGED,
+                          lambda x, y: cb_reader_state_changed(x, y, nfc))
+    nfc.set_mode(nfc.MODE_READER)
+    
     try:
         while True:
             # clear screen
@@ -169,7 +168,7 @@ if __name__ == "__main__":
                     Discord.send(f"illuminance is critical: {SENSOR_DATA.illuminance.get_current()}{SENSOR_DATA.illuminance.unit}")
                     data.last_notified = now
 
-            time.sleep(10)
+            time.sleep(100)
 
     except KeyboardInterrupt:
         # the user ended the program so we absorb the exception
